backend/requirements.txt
# Core backend
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0

# LLM / embeddings
openai==1.14.3
transformers==4.38.0
sentence-transformers==2.6.1

# ML / DL
torch==2.2.0
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1

# Compatibility fixes
protobuf==4.21.0
urllib3==1.26.18
tf-keras==2.15.0

# Other dependencies
httpx
python-dotenv
groq
langchain
sentence-transformers
chromadb
deep-translator
geopy
PyPDF2
sqlalchemy
langchain-community
python-multipart
passlib[bcrypt]
python-jose[cryptography]